niri-config: modularlize configuration Primarily based on top-level config options, with some shared functionality throughout.
