+++ niri-config/src/input.rs
@@ -1,0 +1,262 @@
+use std::str::FromStr;
+
+use miette::miette;
+use smithay::input::keyboard::XkbConfig;
+
+use crate::core::{
+    AccelProfile, ClickMethod, FloatOrInt, ModKey, Percent, ScrollMethod, TapButtonMap,
+};
+
+#[derive(knuffel::Decode, Debug, Default, PartialEq)]
+pub struct Input {
+    #[knuffel(child, default)]
+    pub keyboard: Keyboard,
+    #[knuffel(child, default)]
+    pub touchpad: Touchpad,
+    #[knuffel(child, default)]
+    pub mouse: Mouse,
+    #[knuffel(child, default)]
+    pub trackpoint: Trackpoint,
+    #[knuffel(child, default)]
+    pub trackball: Trackball,
+    #[knuffel(child, default)]
+    pub tablet: Tablet,
+    #[knuffel(child, default)]
+    pub touch: Touch,
+    #[knuffel(child)]
+    pub disable_power_key_handling: bool,
+    #[knuffel(child)]
+    pub warp_mouse_to_focus: Option<WarpMouseToFocus>,
+    #[knuffel(child)]
+    pub focus_follows_mouse: Option<FocusFollowsMouse>,
+    #[knuffel(child)]
+    pub workspace_auto_back_and_forth: bool,
+    #[knuffel(child, unwrap(argument, str))]
+    pub mod_key: Option<ModKey>,
+    #[knuffel(child, unwrap(argument, str))]
+    pub mod_key_nested: Option<ModKey>,
+}
+
+#[derive(knuffel::Decode, Debug, PartialEq, Eq)]
+pub struct Keyboard {
+    #[knuffel(child, default)]
+    pub xkb: Xkb,
+    // The defaults were chosen to match wlroots and sway.
+    #[knuffel(child, unwrap(argument), default = Self::default().repeat_delay)]
+    pub repeat_delay: u16,
+    #[knuffel(child, unwrap(argument), default = Self::default().repeat_rate)]
+    pub repeat_rate: u8,
+    #[knuffel(child, unwrap(argument), default)]
+    pub track_layout: TrackLayout,
+    #[knuffel(child)]
+    pub numlock: bool,
+}
+
+impl Default for Keyboard {
+    fn default() -> Self {
+        Self {
+            xkb: Default::default(),
+            repeat_delay: 600,
+            repeat_rate: 25,
+            track_layout: Default::default(),
+            numlock: Default::default(),
+        }
+    }
+}
+
+#[derive(knuffel::Decode, Debug, Default, PartialEq, Eq, Clone)]
+pub struct Xkb {
+    #[knuffel(child, unwrap(argument), default)]
+    pub rules: String,
+    #[knuffel(child, unwrap(argument), default)]
+    pub model: String,
+    #[knuffel(child, unwrap(argument), default)]
+    pub layout: String,
+    #[knuffel(child, unwrap(argument), default)]
+    pub variant: String,
+    #[knuffel(child, unwrap(argument))]
+    pub options: Option<String>,
+    #[knuffel(child, unwrap(argument))]
+    pub file: Option<String>,
+}
+
+impl Xkb {
+    pub fn to_xkb_config(&self) -> XkbConfig {
+        XkbConfig {
+            rules: &self.rules,
+            model: &self.model,
+            layout: &self.layout,
+            variant: &self.variant,
+            options: self.options.clone(),
+        }
+    }
+}
+
+#[derive(knuffel::DecodeScalar, Debug, Default, PartialEq, Eq)]
+pub enum TrackLayout {
+    /// The layout change is global.
+    #[default]
+    Global,
+    /// The layout change is window local.
+    Window,
+}
+
+#[derive(knuffel::Decode, Debug, Default, PartialEq)]
+pub struct Touchpad {
+    #[knuffel(child)]
+    pub off: bool,
+    #[knuffel(child)]
+    pub tap: bool,
+    #[knuffel(child)]
+    pub dwt: bool,
+    #[knuffel(child)]
+    pub dwtp: bool,
+    #[knuffel(child, unwrap(argument))]
+    pub drag: Option<bool>,
+    #[knuffel(child)]
+    pub drag_lock: bool,
+    #[knuffel(child)]
+    pub natural_scroll: bool,
+    #[knuffel(child, unwrap(argument, str))]
+    pub click_method: Option<ClickMethod>,
+    #[knuffel(child, unwrap(argument), default)]
+    pub accel_speed: FloatOrInt<-1, 1>,
+    #[knuffel(child, unwrap(argument, str))]
+    pub accel_profile: Option<AccelProfile>,
+    #[knuffel(child, unwrap(argument, str))]
+    pub scroll_method: Option<ScrollMethod>,
+    #[knuffel(child, unwrap(argument))]
+    pub scroll_button: Option<u32>,
+    #[knuffel(child)]
+    pub scroll_button_lock: bool,
+    #[knuffel(child, unwrap(argument, str))]
+    pub tap_button_map: Option<TapButtonMap>,
+    #[knuffel(child)]
+    pub left_handed: bool,
+    #[knuffel(child)]
+    pub disabled_on_external_mouse: bool,
+    #[knuffel(child)]
+    pub middle_emulation: bool,
+    #[knuffel(child, unwrap(argument))]
+    pub scroll_factor: Option<FloatOrInt<0, 100>>,
+}
+
+#[derive(knuffel::Decode, Debug, Default, PartialEq)]
+pub struct Mouse {
+    #[knuffel(child)]
+    pub off: bool,
+    #[knuffel(child)]
+    pub natural_scroll: bool,
+    #[knuffel(child, unwrap(argument), default)]
+    pub accel_speed: FloatOrInt<-1, 1>,
+    #[knuffel(child, unwrap(argument, str))]
+    pub accel_profile: Option<AccelProfile>,
+    #[knuffel(child, unwrap(argument, str))]
+    pub scroll_method: Option<ScrollMethod>,
+    #[knuffel(child, unwrap(argument))]
+    pub scroll_button: Option<u32>,
+    #[knuffel(child)]
+    pub scroll_button_lock: bool,
+    #[knuffel(child)]
+    pub left_handed: bool,
+    #[knuffel(child)]
+    pub middle_emulation: bool,
+    #[knuffel(child, unwrap(argument))]
+    pub scroll_factor: Option<FloatOrInt<0, 100>>,
+}
+
+#[derive(knuffel::Decode, Debug, Default, PartialEq)]
+pub struct Trackpoint {
+    #[knuffel(child)]
+    pub off: bool,
+    #[knuffel(child)]
+    pub natural_scroll: bool,
+    #[knuffel(child, unwrap(argument), default)]
+    pub accel_speed: FloatOrInt<-1, 1>,
+    #[knuffel(child, unwrap(argument, str))]
+    pub accel_profile: Option<AccelProfile>,
+    #[knuffel(child, unwrap(argument, str))]
+    pub scroll_method: Option<ScrollMethod>,
+    #[knuffel(child, unwrap(argument))]
+    pub scroll_button: Option<u32>,
+    #[knuffel(child)]
+    pub scroll_button_lock: bool,
+    #[knuffel(child)]
+    pub left_handed: bool,
+    #[knuffel(child)]
+    pub middle_emulation: bool,
+}
+
+#[derive(knuffel::Decode, Debug, Default, PartialEq)]
+pub struct Trackball {
+    #[knuffel(child)]
+    pub off: bool,
+    #[knuffel(child)]
+    pub natural_scroll: bool,
+    #[knuffel(child, unwrap(argument), default)]
+    pub accel_speed: FloatOrInt<-1, 1>,
+    #[knuffel(child, unwrap(argument, str))]
+    pub accel_profile: Option<AccelProfile>,
+    #[knuffel(child, unwrap(argument, str))]
+    pub scroll_method: Option<ScrollMethod>,
+    #[knuffel(child, unwrap(argument))]
+    pub scroll_button: Option<u32>,
+    #[knuffel(child)]
+    pub scroll_button_lock: bool,
+    #[knuffel(child)]
+    pub left_handed: bool,
+    #[knuffel(child)]
+    pub middle_emulation: bool,
+}
+
+#[derive(knuffel::Decode, Debug, Default, PartialEq)]
+pub struct Tablet {
+    #[knuffel(child)]
+    pub off: bool,
+    #[knuffel(child, unwrap(arguments))]
+    pub calibration_matrix: Option<Vec<f32>>,
+    #[knuffel(child, unwrap(argument))]
+    pub map_to_output: Option<String>,
+    #[knuffel(child)]
+    pub left_handed: bool,
+}
+
+#[derive(knuffel::Decode, Debug, Default, PartialEq)]
+pub struct Touch {
+    #[knuffel(child)]
+    pub off: bool,
+    #[knuffel(child, unwrap(argument))]
+    pub map_to_output: Option<String>,
+}
+
+#[derive(knuffel::Decode, Debug, Clone, Copy, PartialEq)]
+pub struct FocusFollowsMouse {
+    #[knuffel(property, str)]
+    pub max_scroll_amount: Option<Percent>,
+}
+
+#[derive(knuffel::Decode, Debug, PartialEq, Eq, Clone, Copy)]
+pub struct WarpMouseToFocus {
+    #[knuffel(property, str)]
+    pub mode: Option<WarpMouseToFocusMode>,
+}
+
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+pub enum WarpMouseToFocusMode {
+    CenterXy,
+    CenterXyAlways,
+}
+
+impl FromStr for WarpMouseToFocusMode {
+    type Err = miette::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "center-xy" => Ok(Self::CenterXy),
+            "center-xy-always" => Ok(Self::CenterXyAlways),
+            _ => Err(miette!(
+                r#"invalid mode for warp-mouse-to-focus, can be "center-xy" or "center-xy-always" (or leave unset for separate centering)"#
+            )),
+        }
+    }
+}
