+++ niri-config/src/rules.rs
@@ -1,0 +1,101 @@
+use niri_ipc::ColumnDisplay;
+
+use crate::appearance::{BorderRule, ShadowRule, TabIndicatorRule};
+use crate::core::FloatOrInt;
+use crate::layout::{CornerRadius, DefaultPresetSize, FloatingPosition};
+use crate::utils::RegexEq;
+
+#[derive(knuffel::Decode, Debug, Default, Clone, PartialEq)]
+pub struct WindowRule {
+    #[knuffel(children(name = "match"))]
+    pub matches: Vec<Match>,
+    #[knuffel(children(name = "exclude"))]
+    pub excludes: Vec<Match>,
+
+    // Rules applied at initial configure.
+    #[knuffel(child)]
+    pub default_column_width: Option<DefaultPresetSize>,
+    #[knuffel(child)]
+    pub default_window_height: Option<DefaultPresetSize>,
+    #[knuffel(child, unwrap(argument))]
+    pub open_on_output: Option<String>,
+    #[knuffel(child, unwrap(argument))]
+    pub open_on_workspace: Option<String>,
+    #[knuffel(child, unwrap(argument))]
+    pub open_maximized: Option<bool>,
+    #[knuffel(child, unwrap(argument))]
+    pub open_fullscreen: Option<bool>,
+    #[knuffel(child, unwrap(argument))]
+    pub open_floating: Option<bool>,
+    #[knuffel(child, unwrap(argument))]
+    pub open_focused: Option<bool>,
+
+    // Rules applied dynamically.
+    #[knuffel(child, unwrap(argument))]
+    pub min_width: Option<u16>,
+    #[knuffel(child, unwrap(argument))]
+    pub min_height: Option<u16>,
+    #[knuffel(child, unwrap(argument))]
+    pub max_width: Option<u16>,
+    #[knuffel(child, unwrap(argument))]
+    pub max_height: Option<u16>,
+
+    #[knuffel(child, default)]
+    pub focus_ring: BorderRule,
+    #[knuffel(child, default)]
+    pub border: BorderRule,
+    #[knuffel(child, default)]
+    pub shadow: ShadowRule,
+    #[knuffel(child, default)]
+    pub tab_indicator: TabIndicatorRule,
+    #[knuffel(child, unwrap(argument))]
+    pub draw_border_with_background: Option<bool>,
+    #[knuffel(child, unwrap(argument))]
+    pub opacity: Option<f32>,
+    #[knuffel(child)]
+    pub geometry_corner_radius: Option<CornerRadius>,
+    #[knuffel(child, unwrap(argument))]
+    pub clip_to_geometry: Option<bool>,
+    #[knuffel(child, unwrap(argument))]
+    pub baba_is_float: Option<bool>,
+    #[knuffel(child, unwrap(argument))]
+    pub block_out_from: Option<BlockOutFrom>,
+    #[knuffel(child, unwrap(argument))]
+    pub variable_refresh_rate: Option<bool>,
+    #[knuffel(child, unwrap(argument, str))]
+    pub default_column_display: Option<ColumnDisplay>,
+    #[knuffel(child)]
+    pub default_floating_position: Option<FloatingPosition>,
+    #[knuffel(child, unwrap(argument))]
+    pub scroll_factor: Option<FloatOrInt<0, 100>>,
+    #[knuffel(child, unwrap(argument))]
+    pub tiled_state: Option<bool>,
+}
+
+#[derive(knuffel::Decode, Debug, Default, Clone, PartialEq)]
+pub struct Match {
+    #[knuffel(property, str)]
+    pub app_id: Option<RegexEq>,
+    #[knuffel(property, str)]
+    pub title: Option<RegexEq>,
+    #[knuffel(property)]
+    pub is_active: Option<bool>,
+    #[knuffel(property)]
+    pub is_focused: Option<bool>,
+    #[knuffel(property)]
+    pub is_active_in_column: Option<bool>,
+    #[knuffel(property)]
+    pub is_floating: Option<bool>,
+    #[knuffel(property)]
+    pub is_window_cast_target: Option<bool>,
+    #[knuffel(property)]
+    pub is_urgent: Option<bool>,
+    #[knuffel(property)]
+    pub at_startup: Option<bool>,
+}
+
+#[derive(knuffel::DecodeScalar, Debug, Clone, Copy, PartialEq, Eq)]
+pub enum BlockOutFrom {
+    Screencast,
+    ScreenCapture,
+}
