+++ niri-config/src/startup.rs
@@ -1,0 +1,70 @@
+#[derive(knuffel::Decode, Debug, Clone, PartialEq, Eq)]
+pub struct SpawnAtStartup {
+    #[knuffel(arguments)]
+    pub command: Vec<String>,
+}
+
+#[derive(knuffel::Decode, Debug, PartialEq)]
+pub struct Cursor {
+    #[knuffel(child, unwrap(argument), default = String::from("default"))]
+    pub xcursor_theme: String,
+    #[knuffel(child, unwrap(argument), default = 24)]
+    pub xcursor_size: u8,
+    #[knuffel(child)]
+    pub hide_when_typing: bool,
+    #[knuffel(child, unwrap(argument))]
+    pub hide_after_inactive_ms: Option<u32>,
+}
+
+impl Default for Cursor {
+    fn default() -> Self {
+        Self {
+            xcursor_theme: String::from("default"),
+            xcursor_size: 24,
+            hide_when_typing: false,
+            hide_after_inactive_ms: None,
+        }
+    }
+}
+
+#[derive(knuffel::Decode, Debug, Default, Clone, Copy, PartialEq, Eq)]
+pub struct HotkeyOverlay {
+    #[knuffel(child)]
+    pub skip_at_startup: bool,
+    #[knuffel(child)]
+    pub hide_not_bound: bool,
+}
+
+#[derive(knuffel::Decode, Debug, Default, Clone, Copy, PartialEq, Eq)]
+pub struct Clipboard {
+    #[knuffel(child)]
+    pub disable_primary: bool,
+}
+
+#[derive(knuffel::Decode, Debug, Default, Clone, PartialEq, Eq)]
+pub struct Environment(#[knuffel(children)] pub Vec<EnvironmentVariable>);
+
+#[derive(knuffel::Decode, Debug, Clone, PartialEq, Eq)]
+pub struct EnvironmentVariable {
+    #[knuffel(node_name)]
+    pub name: String,
+    #[knuffel(argument)]
+    pub value: Option<String>,
+}
+
+#[derive(knuffel::Decode, Debug, Clone, PartialEq, Eq)]
+pub struct XwaylandSatellite {
+    #[knuffel(child)]
+    pub off: bool,
+    #[knuffel(child, unwrap(argument), default = Self::default().path)]
+    pub path: String,
+}
+
+impl Default for XwaylandSatellite {
+    fn default() -> Self {
+        Self {
+            off: false,
+            path: String::from("xwayland-satellite"),
+        }
+    }
+}
