+++ niri-config/src/utils.rs
@@ -1,5 +1,6 @@
 use std::str::FromStr;
 
+use knuffel::errors::DecodeError;
 use regex::Regex;
 
 /// `Regex` that implements `PartialEq` by its string form.
@@ -19,5 +20,36 @@ impl FromStr for RegexEq {
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         Regex::from_str(s).map(Self)
+    }
+}
+
+pub fn expect_only_children<S>(
+    node: &knuffel::ast::SpannedNode<S>,
+    ctx: &mut knuffel::decode::Context<S>,
+) where
+    S: knuffel::traits::ErrorSpan,
+{
+    if let Some(type_name) = &node.type_name {
+        ctx.emit_error(DecodeError::unexpected(
+            type_name,
+            "type name",
+            "no type name expected for this node",
+        ));
+    }
+
+    for val in node.arguments.iter() {
+        ctx.emit_error(DecodeError::unexpected(
+            &val.literal,
+            "argument",
+            "no arguments expected for this node",
+        ))
+    }
+
+    for name in node.properties.keys() {
+        ctx.emit_error(DecodeError::unexpected(
+            name,
+            "property",
+            "no properties expected for this node",
+        ))
     }
 }
